레시피 조회수 , 레시피 정렬 추가

+++ src/main/java/com/example/springjwt/Mypage/LikeRecipeRepository.java
@@ -11,7 +11,7 @@ public interface LikeRecipeRepository extends JpaRepository<LikeRecipe, Long> {
     boolean existsByUserAndRecipe(UserEntity user, Recipe recipe);
     Optional<LikeRecipe> findByUserAndRecipe(UserEntity user, Recipe recipe);
     List<LikeRecipe> findByUser(UserEntity user);
-    Long countByRecipe(Recipe recipe);
+    int countByRecipe(Recipe recipe);
 
     void delete(LikeRecipe likeRecipe);
 }
